backend/requirements.txt
fastapi
uvicorn[standard]
websockets
google-cloud-firestore
google-cloud-pubsub
google-cloud-aiplatform
google-cloud-monitoring
google-cloud-logging
google-cloud-run
google-auth
langchain
langchain-google-vertexai
langgraph
faiss-cpu
numpy
pydantic
python-dotenv
httpx
python-multipart
sentence-transformers
scikit-learn
slowapi
